Rename matomo.js to run-vm.client.js
Remove plugin settings from the config as everything was already loading automatically.
Adblockers seem to work when the matomo plugin isn't called matomo.
